Extract JSON queries to a separate class
